Removed toUint8Array and added more toBuffer documentation instead.
