+++ lib/input.js
@@ -32,7 +32,7 @@ function _createInputDescriptor (input, inputOptions, containerOptions) {
     // Buffer
     inputDescriptor.buffer = input;
   } else if (is.uint8Array(input)) {
-    // Uint8Array
+    // Uint8Array or Uint8ClampedArray
     inputDescriptor.buffer = Buffer.from(input.buffer);
   } else if (is.plainObject(input) && !is.defined(inputOptions)) {
     // Plain Object descriptor, e.g. create
