+++ lib/output.js
@@ -104,6 +104,20 @@ function toFile (fileOut, callback) {
  *   .then(({ data, info }) => { ... })
  *   .catch(err => { ... });
  *
+ * @example
+ * const data = await sharp('my-image.jpg')
+ *   // output the raw pixels
+ *   .raw()
+ *   .toBuffer();
+ * 
+ * // create a more type safe way to work with the raw pixel data
+ * // this will not copy the data, instead it will change `data`s underlying ArrayBuffer
+ * // so `data` and `pixelArray` point to the same memory location
+ * const pixelArray = new Uint8ClampedArray(data.buffer);
+ * 
+ * // When you are done changing the pixelArray, sharp takes the `pixelArray` as an input
+ * await sharp(pixelArray).toFile('my-changed-image.jpg');
+ *
  * @param {Object} [options]
  * @param {boolean} [options.resolveWithObject] Resolve the Promise with an Object containing `data` and `info` properties instead of resolving only with `data`.
  * @param {Function} [callback]
@@ -116,48 +130,6 @@ function toBuffer (options, callback) {
     this.options.resolveWithObject = false;
   }
   return this._pipeline(is.fn(options) ? options : callback);
-}
-
-/**
- * Same as toBuffer, but writes output to a Uint8Array.
- *
- * @example
- * sharp(input)
- *   .toUint8Array((err, data, info) => { ... });
- *
- * @example
- * sharp(input)
- *   .toUint8Array()
- *   .then(data => { ... })
- *   .catch(err => { ... });
- *
- * @example
- * sharp(input)
- *   .toUint8Array({ resolveWithObject: true })
- *   .then(({ data, info }) => { ... })
- *   .catch(err => { ... });
- *
- * @param {Object|Function} [optionsOrCallback] The resolveWithObject option if you want to the promise to include info.
- * @param {boolean} [options.resolveWithObject] Resolve the Promise with an Object containing `data` and `info` properties instead of resolving only with `data`.
- * @returns {Promise<Uint8Array>} - when no callback is provided
- */
-function toUint8Array (optionsOrCallback) {
-  // callback mode
-  if (is.fn(optionsOrCallback)) {
-    // wrap callback with function that converts to Uint8Array
-    return this.toBuffer((err, data, info) => optionsOrCallback(err, new Uint8Array(data.buffer), info));
-  }
-  return this.toBuffer(optionsOrCallback).then((res) => {
-    if (optionsOrCallback && optionsOrCallback.resolveWithObject) {
-      return {
-        // creating a new Uint8Array with the ArrayBuffer of a Buffer means the Uint8Array with share underlying memory
-        data: new Uint8Array(res.data.buffer),
-        info: res.info
-      };
-    } else {
-      return new Uint8Array(res.buffer);
-    }
-  });
 }
 
 /**
@@ -948,7 +920,6 @@ module.exports = function (Sharp) {
     // Public
     toFile,
     toBuffer,
-    toUint8Array,
     withMetadata,
     toFormat,
     jpeg,
